+++ src/Libraries/SmartStore.Services/DataExchange/Export/ExportXmlHelper.cs
@@ -202,7 +202,6 @@ namespace SmartStore.Services.DataExchange.Export
 			_writer.Write("CreatedOnUtc", entity.CreatedOnUtc.ToString(_culture));
 			_writer.Write("UpdatedOnUtc", entity.UpdatedOnUtc.ToString(_culture));
 			_writer.Write("DomainEndings", entity.DomainEndings);
-			_writer.Write("RoundingMethod", ((int)entity.RoundingMethod).ToString());
 
 			WriteLocalized(currency);