+++ src/Presentation/SmartStore.Web/Administration/Controllers/CurrencyController.cs
@@ -6,13 +6,14 @@ using System.Web.Mvc;
 using SmartStore.Admin.Models.Directory;
 using SmartStore.Core.Domain.Directory;
 using SmartStore.Core.Domain.Stores;
+using SmartStore.Core.Plugins;
 using SmartStore.Services;
 using SmartStore.Services.Directory;
 using SmartStore.Services.Helpers;
 using SmartStore.Services.Localization;
+using SmartStore.Services.Payments;
 using SmartStore.Services.Security;
 using SmartStore.Services.Stores;
-using SmartStore.Web.Framework;
 using SmartStore.Web.Framework.Controllers;
 using SmartStore.Web.Framework.Filters;
 using SmartStore.Web.Framework.Plugins;
@@ -21,7 +22,7 @@ using Telerik.Web.Mvc;
 
 namespace SmartStore.Admin.Controllers
 {
-	[AdminAuthorize]
+    [AdminAuthorize]
     public class CurrencyController :  AdminControllerBase
     {
         #region Fields
@@ -34,28 +35,32 @@ namespace SmartStore.Admin.Controllers
 		private readonly IStoreMappingService _storeMappingService;
 		private readonly PluginMediator _pluginMediator;
 		private readonly ICommonServices _services;
+        private readonly IPaymentService _paymentService;
 
         #endregion
 
         #region Constructors
 
-        public CurrencyController(ICurrencyService currencyService, 
+        public CurrencyController(
+            ICurrencyService currencyService, 
             CurrencySettings currencySettings,
             IDateTimeHelper dateTimeHelper,
             ILocalizedEntityService localizedEntityService,
 			ILanguageService languageService,
             IStoreMappingService storeMappingService,
 			PluginMediator pluginMediator,
-			ICommonServices services)
+			ICommonServices services,
+            IPaymentService paymentService)
         {
-            this._currencyService = currencyService;
-            this._currencySettings = currencySettings;
-            this._dateTimeHelper = dateTimeHelper;
-            this._localizedEntityService = localizedEntityService;
-            this._languageService = languageService;
-			this._storeMappingService = storeMappingService;
-			this._pluginMediator = pluginMediator;
-			this._services = services;
+            _currencyService = currencyService;
+            _currencySettings = currencySettings;
+            _dateTimeHelper = dateTimeHelper;
+            _localizedEntityService = localizedEntityService;
+            _languageService = languageService;
+			_storeMappingService = storeMappingService;
+			_pluginMediator = pluginMediator;
+			_services = services;
+            _paymentService = paymentService;
         }
         
         #endregion
@@ -80,9 +85,22 @@ namespace SmartStore.Admin.Controllers
 				throw new ArgumentNullException("model");
 
 			var allStores = _services.StoreService.GetAllStores();
+            var paymentMethods = _paymentService.GetAllPaymentMethods();
+            var paymentProviders = _paymentService.LoadAllPaymentMethods().ToDictionarySafe(x => x.Metadata.SystemName);
 
-			model.AvailableStores = allStores.Select(s => s.ToModel()).ToList();
-			model.AvailableRoundingMethods = CurrencyRoundingMethod.Up005.ToSelectList(false).ToList();
+            model.AvailableStores = allStores.Select(s => s.ToModel()).ToList();
+
+            foreach (var paymentMethod in paymentMethods.Where(x => x.RoundOrderTotal))
+            {
+                string friendlyName = null;
+                Provider<IPaymentMethod> provider;
+                if (paymentProviders.TryGetValue(paymentMethod.PaymentMethodSystemName, out provider))
+                {
+                    friendlyName = _pluginMediator.GetLocalizedFriendlyName(provider.Metadata);
+                }
+
+                model.RoundOrderTotalPaymentMethods[paymentMethod.PaymentMethodSystemName] = friendlyName ?? paymentMethod.PaymentMethodSystemName;
+            }
 
 			if (currency != null)
 			{
@@ -118,11 +136,6 @@ namespace SmartStore.Admin.Controllers
 
 			model.IsPrimaryStoreCurrency = store.PrimaryStoreCurrencyId == model.Id;
 			model.IsPrimaryExchangeRateCurrency = store.PrimaryExchangeRateCurrencyId == model.Id;
-
-			if (model.RoundingMethod.HasValue)
-			{
-				model.RoundingMethodString = model.RoundingMethod.Value.GetLocalizedEnum(Services.Localization, Services.WorkContext);
-			}
 
 			return model;
 		}
@@ -306,13 +319,13 @@ namespace SmartStore.Admin.Controllers
                 return AccessDeniedView();
 
             var model = new CurrencyModel();
-            //locales
             AddLocales(_languageService, model.Locales);
-			//Stores
 			PrepareCurrencyModel(model, null, false);
-            //default values
+
+            // Default values
             model.Published = true;
             model.Rate = 1;
+
             return View(model);
         }
 
@@ -328,19 +341,15 @@ namespace SmartStore.Admin.Controllers
                 
 				_currencyService.InsertCurrency(currency);
                 
-				//locales
                 UpdateLocales(currency, model);
 				
-				//Stores
-				_storeMappingService.SaveStoreMappings<Currency>(currency, model.SelectedStoreIds);
+				_storeMappingService.SaveStoreMappings(currency, model.SelectedStoreIds);
 
                 NotifySuccess(_services.Localization.GetResource("Admin.Configuration.Currencies.Added"));
                 return continueEditing ? RedirectToAction("Edit", new { id = currency.Id }) : RedirectToAction("List");
             }
 
-            //If we got this far, something failed, redisplay form
-
-			//Stores
+            // If we got this far, something failed, redisplay form
 			PrepareCurrencyModel(model, null, true);
 
             return View(model);
@@ -358,7 +367,6 @@ namespace SmartStore.Admin.Controllers
             var model = currency.ToModel();
             model.CreatedOn = _dateTimeHelper.ConvertToUserTime(currency.CreatedOnUtc, DateTimeKind.Utc);
             
-			//locales
             AddLocales(_languageService, model.Locales, (locale, languageId) =>
             {
                 locale.Name = currency.GetLocalized(x => x.Name, languageId, false, false);
@@ -367,13 +375,16 @@ namespace SmartStore.Admin.Controllers
 			foreach (var ending in model.DomainEndings.SplitSafe(","))
 			{
 				var item = model.AvailableDomainEndings.FirstOrDefault(x => x.Value.IsCaseInsensitiveEqual(ending));
-				if (item == null)
-					model.AvailableDomainEndings.Add(new SelectListItem() { Text = ending, Value = ending, Selected = true });
-				else
-					item.Selected = true;
+                if (item == null)
+                {
+                    model.AvailableDomainEndings.Add(new SelectListItem { Text = ending, Value = ending, Selected = true });
+                }
+                else
+                {
+                    item.Selected = true;
+                }
 			}
 
-			//Stores
 			PrepareCurrencyModel(model, currency, false);
 
             return View(model);
@@ -397,10 +408,8 @@ namespace SmartStore.Admin.Controllers
 				{
 					_currencyService.UpdateCurrency(currency);
                 
-					//locales
 					UpdateLocales(currency, model);
 				
-					//Stores
 					_storeMappingService.SaveStoreMappings<Currency>(currency, model.SelectedStoreIds);
 
 					NotifySuccess(_services.Localization.GetResource("Admin.Configuration.Currencies.Updated"));
@@ -408,10 +417,9 @@ namespace SmartStore.Admin.Controllers
 				}
             }
 
-            //If we got this far, something failed, redisplay form
+            // If we got this far, something failed, redisplay form
             model.CreatedOn = _dateTimeHelper.ConvertToUserTime(currency.CreatedOnUtc, DateTimeKind.Utc);
 
-			//Stores
 			PrepareCurrencyModel(model, currency, true);
 
             return View(model);