+++ src/Tests/SmartStore.Services.Tests/Catalog/PriceCalculationServiceTests.cs
@@ -6,22 +6,24 @@ using Rhino.Mocks;
 using SmartStore.Core;
 using SmartStore.Core.Domain.Catalog;
 using SmartStore.Core.Domain.Customers;
+using SmartStore.Core.Domain.Directory;
 using SmartStore.Core.Domain.Discounts;
 using SmartStore.Core.Domain.Orders;
 using SmartStore.Core.Domain.Stores;
+using SmartStore.Core.Domain.Tax;
 using SmartStore.Services.Catalog;
 using SmartStore.Services.Discounts;
 using SmartStore.Services.Media;
 using SmartStore.Services.Tax;
 using SmartStore.Tests;
-using SmartStore.Core.Domain.Tax;
 
 namespace SmartStore.Services.Tests.Catalog
 {
-	[TestFixture]
+    [TestFixture]
     public class PriceCalculationServiceTests : ServiceTest
     {
 		IStoreContext _storeContext;
+        IWorkContext _workContext;
         IDiscountService _discountService;
         ICategoryService _categoryService;
 		IManufacturerService _manufacturerService;
@@ -37,13 +39,22 @@ namespace SmartStore.Services.Tests.Catalog
         CatalogSettings _catalogSettings;
 		TaxSettings _taxSettings;
 		Store _store;
+        Currency _currency;
 
         [SetUp]
         public new void SetUp()
         {
-			_store = new Store() { Id = 1 };
+			_store = new Store { Id = 1 };
 			_storeContext = MockRepository.GenerateMock<IStoreContext>();
 			_storeContext.Expect(x => x.CurrentStore).Return(_store);
+
+            _currency = new Currency { Id = 1 };
+            _workContext = MockRepository.GenerateMock<IWorkContext>();
+            _workContext.Expect(x => x.WorkingCurrency).Return(_currency);
+
+            _services = MockRepository.GenerateMock<ICommonServices>();
+            _services.Expect(x => x.StoreContext).Return(_storeContext);
+            _services.Expect(x => x.WorkContext).Return(_workContext);
 
             _discountService = MockRepository.GenerateMock<IDiscountService>();
 
@@ -55,8 +66,6 @@ namespace SmartStore.Services.Tests.Catalog
 			_productAttributeService = MockRepository.GenerateMock<IProductAttributeService>();
 
 			_downloadService = MockRepository.GenerateMock<IDownloadService>();
-			_services = MockRepository.GenerateMock<ICommonServices>();
-			_services.Expect(x => x.StoreContext).Return(_storeContext);
 			_httpRequestBase = MockRepository.GenerateMock<HttpRequestBase>();
 			_taxService = MockRepository.GenerateMock<ITaxService>();